fix: cast m->index to long long for printk

+++ code/hide_mounts.c
@@ -73,7 +73,7 @@ static int iter_ret(struct kretprobe_instance *ri, struct pt_regs *regs)
         return 0;
     }
     printk(KERN_INFO "hm: iter_ret: count=%zu index=%zu size=%zu\n",
-           m->count, m->index, m->size);
+           m->count, (long long)m->index, m->size);
 
     if (m->count == 0)
         return 0;
